Кнопка логина перенесена в header
